requirements.txt
aniso8601==7.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.4.1
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.1
graphql-server-core==1.1.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
promise==2.2.1
PyJWT==1.7.1
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
SQLAlchemy==1.1.14
urllib3==1.25.7
Werkzeug==0.16.0
